Workaround for disposal bug
